This looks like it is commiting to the android version, but it should be the web version. Hopefully git knows what it is doing.
I fixed a problem with the thymeleaf syntax in the audit panel. Thymeleaf is so simple, and yet seems to need a good deal of getting used to. Then again, I did convert everything in about two hours.  Although, the problems still appear from time to time, but most of them were present in the old way of doing things as well.

With the thymeleaf revisions coming to a close, I feel that I should revisit the to do list that I commited earlier.

  TODO:
    code cleaning
    remove deprecated methods
    custom validators
    thymeleaf
      - make integration tests for thymeleaf
    simplify resource files
    check pagination on address resource
      - run order pagination tests on address resource
    load paging links using aspects

====================================
    custom validators
     - done, except order in state that is not supported.
    	This should be of very little problem, because the dropdown only supports states in the database, but still. It should validate that.

	thymeleaf
		- done, except for the random improvements that I may yet make.
		- - integration tests using thymeleaf methods have been forgotten for now.  Content type is not supported errors seem to be plagueing me and I can not figure out why.  It works correctly in production, just not through the mockup. Actually, I think that I is the post that is messing with it, or maybe the json. Probably the post, I took away all references to json and it stil did not work. No reason to think that the json specification was getting through though. Too much time waisted on this feature already, I should move on. Perhaps stackoverflow.com could help later.

====================================
	bootstrap menu bar
	mobile friendly show order page
	tests with better fixturing and cleanup
		(the current tests need to have a few things added manually, an then removed for the integeration tests to pass)

In conclusion, I feel like I spent too much of today working something that just was not meant to work out. Perhaps tomorrow will be more productive.